Alert Creation

+++ Integratie.Domain/Entities/Alerts/CheckAlert.cs
@@ -10,7 +10,7 @@ namespace Integratie.Domain.Entities.Alerts
     //Basic Alert
     public class CheckAlert : Alert
     {
-        public CheckAlert(SubjectProperty subjectProperty, Operator @operator, double value, Subject subject)
+        public CheckAlert(SubjectProperty subjectProperty, Operator @operator, int value, Subject subject)
         {
             SubjectProperty = subjectProperty;
             Operator = @operator;
@@ -24,7 +24,7 @@ namespace Integratie.Domain.Entities.Alerts
 
         public SubjectProperty SubjectProperty { get; set; }
         public Operator Operator { get; set; }
-        public double Value { get; set; }
+        public int Value { get; set; }
         public Subject Subject { get; set; }
     }
 }
