+++ Integratie.Domain/Entities/Alerts/SentimentAlert.cs
@@ -11,10 +11,10 @@ namespace Integratie.Domain.Entities.Alerts
     {
         public SubjectProperty SubjectProperty { get; set; }
         public Operator Operator { get; set; }
-        public double Value { get; set; }
+        public int Value { get; set; }
         public Subject Subject { get; set; }
 
-        public SentimentAlert(SubjectProperty subjectProperty, Operator @operator, double value, Subject subject)
+        public SentimentAlert(SubjectProperty subjectProperty, Operator @operator, int value, Subject subject)
         {
             SubjectProperty = subjectProperty;
             Operator = @operator;
